LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A6_T2-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Expected an identifier but found '\u000D' instead
S7.3_A6_T2

PASS Expected parsing failure

TEST COMPLETE
